% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixmodel.R
\name{mixmodel}
\alias{as.mixmodel}
\alias{is.mixmodel}
\alias{mixmodel}
\title{mixmodel}
\usage{
mixmodel(family, type, invlink = NULL, dlink = NULL)

is.mixmodel(x)

as.mixmodel(x)
}
\arguments{
\item{family}{a character string representing the exponential family of the
mixture densities. The current supported families are \code{"normal"},
\code{"poisson"}, \code{"binomial"} and \code{"mvnormal"}.}

\item{type}{a character string representing the type of mixture. The current
supported types are \code{"mixef"} for mixtures of exponential families,
\code{"mixglm"} for mixtures of generalised linear models.}

\item{invlink}{a function representing the inverse link function in the
generalised linear model, i.e. the function that maps the linear predictor to
the mean. This is NULL by default, which chooses the canonical inverse link
of the exponential family provided by the \code{family} argument.}

\item{dlink}{a function representing the derivative of the link function in
the generalised linear model, i.e. the derivative of the function that maps
the mean to the linear predictor. This is NULL by default, which chooses the
derivative of the canonical link of the exponential family provided by the
\code{family} argument.}

\item{x}{a list to be coerced or object to be verified.}
}
\value{
Returns an object of class \code{mixmodel}, which is essentially a
list with the following components:
\item{family}{a character string equal to the value provided to the
\code{family} argument.}
\item{type}{a character string equal to the value provided to the \code{type}
argument.}
\item{samp}{a function that samples from the exponential family used for this
mixture model.}
\item{sampmap}{a named vector of character strings that represent the names
of the arguments required by the sampling function. The names correspond to
the names used for these parameters when specifying parameters (e.g. in
\code{\link{mixstream}} or \code{\link{rmixture}}).}
\item{dens}{a function that computes the product of a mixture probability and
the density of the exponential family used for the mixture model. Note that
this density only needs to be correct up to a constant of proportionality.}
\item{suff}{a function that returns the sufficient statistic of the supplied
data given the exponential family of the mixture model.}
\item{upd}{a function that updates the parameter estimates give the
sufficient statistics}
\item{suffdims}{a numeric vector representing the }
}
\description{
Specify the type of model for use in \code{\link{mixstream}}.
}
\section{Specifying custom mixture models}{

}
\seealso{
\code{\link{mixstream}}
}

