% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixsettings.R
\name{mixsettings}
\alias{as.mixsettings}
\alias{is.mixsettings}
\alias{mixsettings}
\title{mixsettings}
\usage{
mixsettings(tours, toursize, rate = 0.6, delay = 20, PRavg = 0,
  keep_history = FALSE, debugit = 0, eta = 0.01, eta1 = NULL,
  eta2 = NULL, rprop = 0)

is.mixsettings(x)

as.mixsettings(x, tours, toursize)
}
\arguments{
\item{tours}{an integer specifying the number of tours to run.}

\item{toursize}{an integer specifying the number of data points in each tour.}

\item{rate}{a double in the open interval (0.5, 1). This represents the
exponent to be used for the learning rate in the online EM algorithm where
the learning rate is of the form gamma = i^(-rate) where i is the current
iteration. Larger values lead to quicker convergence but less stability and
vice-versa for smaller values. The default is 0.6, which is often a suitable
compromise.}

\item{delay}{an integer representing the length of the inhibition phase, i.e.
the number of iterations of the online EM algorithm to run before the
parameters of the mixture model are updated. Having a small inhibition phase
provides a level of robustness for the early parameter estimates. The default
is 20.}

\item{PRavg}{an integer representing the tour to start Polyak-Ruppert
averaging on, which computes a rolling average of the the parameter estimates
from this value onwards. This is useful for reducing oscillatory behaviour
from the parameter estimates. The default is 0, which results in no
Polyak-Ruppert averaging being applied.}

\item{keep_history}{a logical. If \code{TRUE}, the parameter estimates at
each tour are stored and returned as the \code{history} component in the
output of the \code{\link{mixstream}} function. Warning: this setting is not
recommended for a large number of data points due to the large memory
requirement. The default is \code{FALSE}.}

\item{debugit}{an integer representing the iteration at which to pause the
algorithm and open up the R browser. Useful for debugging. The default is 0,
i.e the algorithm is not interrupted by default.}

\item{eta}{a double representing a constant learning rate when using SGD to
estimate the regression coefficients. Defaults to 0.01.}

\item{eta1}{a double representing the first parameter in a decreasing
learning rate of the form eta = eta1 / (1 + eta1 * eta2 * it) when using SGD
to estimate the regression coefficients. Default is \code{NULL}, i.e. not to
use the decreasing learning rate form.}

\item{eta2}{a double representing the second parameter in a decreasing
learning rate of the form eta = eta1 / (1 + eta1 * eta2 * it) when using SGD
to estimate the regression coefficients. The default is \code{NULL}, i.e. not
to use the decreasing learning rate form.}

\item{rprop}{a double representing the proportion to increase or decrease the
learning rate when using SGD. Based on RPROP, the learning rate will be
increased if the sign of the previous score agrees with the sign of the
current score, otherwise the learning rate will be decreased. This ought to
provide faster convergence. Here the score refers to the derivative of the
expectation of the complete log-likelihood with respect to the parameters.
The default is 0, i.e. by default the learning rate won't be increased or
decreased.}

\item{x}{a list to be coerced or object to be verified.}
}
\value{
Both \code{mixsettings} and \code{as.mixsettings} functions return
an object of class \code{mixsettings}, which is a list of all the settings
that can be used by the \code{\link{mixstream}} function.

\code{is.mixsettings} returns a logical, which is \code{TRUE} if \code{x}
is a valid \code{mixsettings} object.
}
\description{
Settings to configure \code{\link{mixstream}}.
}
\details{
Note that the \code{\link{mixstream}} function will call
\code{as.mixsettings} on the value of the settings argument by default and
so it is fine to provide a list instead of a \code{mixsettings} object.
Note also that \code{tours} and \code{toursize} need to be specified when
calling the \code{mixsettings} and \code{as.mixsettings} functions, and by
default \code{\link{mixstream}} sets \code{tours} to \code{n} (when \code{n} is
known) and \code{toursize} to 1.
}
\seealso{
\code{\link{mixstream}}
}

