% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixstream.R
\name{mixstream}
\alias{mixstream}
\title{Online Mixture Modelling}
\usage{
mixstream(data, m, model, params, settings)
}
\arguments{
\item{data}{a matrix or data frame of the variables in the model, where the
first column should be the response variable.}

\item{m}{an integer representing the number of finite mixtures.}

\item{model}{a list or \code{\link{mixmodel}} object specifying the type of
mixture model.}

\item{params}{a list of parameters for \code{mixstream} to start at that
should correspond to \code{model}. See 'Details' for more.}

\item{settings}{a list or \code{\link{mixsettings}} object specifying the settings
for mixstream.}
}
\description{
Online Mixture Modelling
}
\seealso{
\code{\link{mixmodel}} for specifying mixture models,
\code{\link{mixsettings}} for configuring mixstream and
\code{\link{rmixture}} for simulating from mixture models.
}

