% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmixture.R
\name{rmixture}
\alias{rmixture}
\title{Simulating from a mixture model.}
\usage{
rmixture(n, m, model, params, data = NULL, shuffle = TRUE)
}
\arguments{
\item{n}{an integer representing the number of data points to simulate.}

\item{m}{an integer representing the number of mixture components.}

\item{model}{a \code{\link{mixmodel}} object representing the type of mixture
and exponential family to be used.}

\item{params}{a list of parameters that should correspond to \code{model} and
may include the mixing proportions. See 'Details' for more.}

\item{data}{a matrix or data frame representing the design matrix of the
covariates. Only required when simulating mixtures of generalised linear
models.}

\item{shuffle}{logical. If \code{TRUE} then shuffle the output.}
}
\value{
Returns a list containing the following components:
 \item{data}{a data frame, which consists of the simulated mixture model data
 . When the model is of type \code{"mixef"} this will be the simulated random
 variables. For models of type \code{"mixglm"} the data frame will contain
 the response variables in the first column and the covariates in the
 remaining columns.}
 \item{component}{a vector of integers from 1 to \code{m} representing the
 mixture component of the associated row in \code{data}.}
}
\description{
\code{rmixture} is used to simulate from a mixture model by
specifying the type of model (e.g. a mixture of exponential families or
generalised linear models) and the necessary parameters of the model type.
}
\details{
The value provided to the \code{params} argument should be a list
with components representing the parameters of the mixture model given by
\code{model} (see \code{\link{mixmodel}} for more information). The
proportions of each mixture can be specified in the \code{"props"} component
of \code{params}. This should be a numeric vector, where each entry is
between 0 and 1 such that the sum of all entries is 1. If the \code{"props"}
component is missing then it assumed that all mixing propotions are equal.
Similarly, if any proportions are not specified, the remaining proportion
will be evenly distributed amongst the unspecified proportions.
}
\seealso{
\code{\link{mixmodel}}
}

